change ActivityOption.setCapacity(capacity) and getCapacity() to work with primitive type double, and not Object-type Double. ActivityOptionImpl.setCapacity() accessed the given capacity unconditionally, thus resulting in a NullPointerException if someone ever tried to pass a Null-Object to it. As it never worked, I guess it's save to just change it to the primitive type.
git-svn-id: file:///srv/local/data/cleanRepo/trunk@25312 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/ciarif/src/main/java/playground/ciarif/retailers/IO/LinksRetailerReader.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.TreeMap;
+
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
@@ -17,6 +18,7 @@ import org.matsim.core.gbl.Gbl;
 import org.matsim.core.gbl.MatsimRandom;
 import org.matsim.core.network.LinkImpl;
 import org.matsim.core.network.NetworkImpl;
+
 import playground.ciarif.retailers.data.LinkRetailersImpl;
 import playground.ciarif.retailers.data.Retailer;
 import playground.ciarif.retailers.data.Retailers;
@@ -146,7 +148,7 @@ public class LinksRetailerReader
         for (ActivityFacility facility : facilities)
         {
           if (facility.getActivityOptions().get("shopgrocery") != null) {
-            double shopCapacity = ((ActivityOption)facility.getActivityOptions().get("shopgrocery")).getCapacity().doubleValue();
+            double shopCapacity = ((ActivityOption)facility.getActivityOptions().get("shopgrocery")).getCapacity();
             globalCapacity += shopCapacity;
           }
         }
@@ -205,7 +207,7 @@ public class LinksRetailerReader
         int numberShops = 0;
         for (ActivityFacility facility : facilities) {
           if (facility.getActivityOptions().get("shopgrocery") != null) {
-            double shopCapacity = ((ActivityOption)facility.getActivityOptions().get("shopgrocery")).getCapacity().doubleValue();
+            double shopCapacity = ((ActivityOption)facility.getActivityOptions().get("shopgrocery")).getCapacity();
             globalCapacity += shopCapacity;
             numberShops = numberShops+1;
           }
